+++ src/main/java/de/samply/teiler/SearchController.java
@@ -3,6 +3,7 @@ package de.samply.teiler;
 import de.samply.converter.Format;
 import de.samply.core.TeilerCore;
 import de.samply.core.TeilerCoreException;
+import de.samply.core.TeilerParameters;
 import de.samply.utils.ProjectVersion;
 import java.nio.file.Path;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +12,8 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Flux;
@@ -37,21 +40,26 @@ public class SearchController {
   }
 
   @GetMapping(value = TeilerConst.RESPONSE, produces = MediaType.APPLICATION_NDJSON_VALUE)
-  public Flux<Path> getResponse(@RequestParam(TeilerConst.QUERY_ID) String queryId) {
+  public Flux<Path> getResponse() {
     //TODO
     return null;
   }
 
   @GetMapping(value = TeilerConst.RETRIEVE_QUERY, produces = MediaType.APPLICATION_NDJSON_VALUE)
   public Flux<Path> retrieveQuery(
-      @RequestParam(TeilerConst.QUERY) String query,
-      @RequestParam(TeilerConst.SOURCE_ID) String sourceId,
-      @RequestParam(TeilerConst.TEMPLATE_ID) String templateId,
-      @RequestParam(TeilerConst.QUERY_FORMAT) Format queryFormat,
-      @RequestParam(TeilerConst.OUTPUT_FORMAT) Format outputFormat
+      @RequestParam(name = TeilerConst.QUERY_ID, required = false) String queryId,
+      @RequestParam(name = TeilerConst.QUERY, required = false) String query,
+      @RequestParam(name = TeilerConst.SOURCE_ID) String sourceId,
+      @RequestParam(name = TeilerConst.QUERY_FORMAT) Format queryFormat,
+      @RequestParam(name = TeilerConst.OUTPUT_FORMAT) Format outputFormat,
+      @RequestParam(name = TeilerConst.TEMPLATE_ID, required = false) String templateId,
+      @RequestHeader(name = "Content-Type", required = false) String contentType,
+      @RequestBody(required = false) String template
   ) {
     try {
-      return teilerCore.retrieveByQuery(sourceId, query, queryFormat, outputFormat, templateId);
+      return teilerCore.retrieveQuery(
+          new TeilerParameters(queryId, query, sourceId, templateId, template, contentType,
+              queryFormat, outputFormat));
     } catch (TeilerCoreException e) {
       //TODO
       throw new RuntimeException(e);